pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "seguimiento_parlamentario"
version = "0.1.0"
description = "A simple Python project for extracting and processing data from the chilean Parliament."
authors = [
    { name = "Diego Acevedo", email = "dacevedo@dcc.uchile.cl" }
]

requires-python = ">=3.10"

dependencies = [
    "babel==2.17.0",
    "celery==5.4.0",
    "fastapi==0.115.13",
    "ffmpeg-python==0.2.0",
    "google-cloud-firestore==2.21.0",
    "google-cloud-tasks==2.19.2",
    "jinja2==3.1.6",
    "markdown==3.8.2",
    "markdown_pdf==1.7",
    "numpy==2.1.3",
    "openai==1.78.1",
    "pinecone==7.0.2",
    "pymongo==4.11.3",
    "pytz==2025.2",
    "requests==2.32.3",
    "selenium==4.30.0",
    "tiktoken==0.9.0",
    "uvicorn==0.34.3",
    "yt-dlp==2025.6.9",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "black",
    "flake8"
]

test = [
    "pytest",
    "pytest-cov"
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools]
py-modules = ["setup"]
